Change: Extract ErrorMarker from useFormValidation hook module
Hooks should not contain JSX code and should be plain JS only. Therefore
extract the ErrorMarker component from the useFormValidation module and
use `js` file suffix for the module.
