+++ src/web/components/form/useFormValidation.js
@@ -3,33 +3,11 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-
-import React, {useState, useEffect, useCallback} from 'react';
-
-import styled from 'styled-components';
+import {useState, useEffect, useCallback} from 'react';
 
 import {_, _l} from 'gmp/locale/lang';
 
 import {hasValue, isDefined, isFunction} from 'gmp/utils/identity';
-
-import {styledExcludeProps} from 'web/utils/styledConfig';
-import Theme from 'web/utils/theme';
-
-const StyledMarker = styledExcludeProps(styled.div, ['isVisible'])`
-  color: ${Theme.darkRed};
-  color: ${props => props.color};
-  font-weight: bold;
-  font-size: 19px;
-  padding-bottom: 1px;
-  padding-left: 4px;
-  display: ${props => (props.isVisible ? 'inline' : 'none')};
-`;
-
-export const Marker = props => (
-  <StyledMarker {...props} data-testid="error-marker">
-    ×
-  </StyledMarker>
-);
 
 const useFormValidation = (
   validationRules,
